requirements.txt
flask
scikit-learn
pandas
pillow
pytesseract